Refactor data models and update API interactions
- Introduce new data models from `SurfTimer.Shared` for consistency across `Plugin` and `API`
- Remove overlapping models
- Update `AccessService`s to match new structures (MySQL **NOT** tested)
- `ApiMethod` includes custom JSON serialization option for `ReplayFramesStringConverter` to match with `API`

+++ src/ST-UTILS/Extensions.cs
@@ -60,7 +60,7 @@ unsafe static class Extensions
     /// </summary>
     /// <param name="tier">Map Tier up to 8</param>
     /// <returns>Appropriate ChatColor value for the Tier</returns>
-    public static char GetTierColor(int tier)
+    public static char GetTierColor(short tier)
     {
         return tier switch
         {
